Line iterators
Start moving functionality into iterators.

These split items into display lines. Each display line is marked with
whether it must be displayed or not (this is the same as "is related to
the selected items, that is lines of text and the surrounding
separators")

This approach allows composing features instead of having a complicated
main loop or a lot of independent implementations. I plan on extending
it to have viewports, gutters (for highlight or informational symbols),
eventually composing a feature pipeline out of iterators to do most of
the work in one pass if possible (if not, there's always an iterator
that consumes it's iterable then loops over the resulting vec :X).